Restructure packaging for entercash
COAND-497
