+++ entercash/src/main/java/com/adyen/checkout/entercash/EntercashComponent.kt
@@ -14,6 +14,7 @@ import com.adyen.checkout.components.base.ComponentEventHandler
 import com.adyen.checkout.components.model.payments.request.EntercashPaymentMethod
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.entercash.EntercashComponent.Companion.PROVIDER
+import com.adyen.checkout.entercash.internal.provider.EntercashComponentProvider
 import com.adyen.checkout.issuerlist.IssuerListComponent
 import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
